requirements.txt
requests==2.31.0
python-dotenv==1.0.0
web3==6.20.1
eth-account==0.11.3
schedule==1.2.2
pandas==2.0.3
sqlalchemy==2.0.23
apscheduler==3.10.4
